Add files via upload
Simply, I was learning how to scrape using BS4, I made these 4 projects to share with you, I have inserted comments also so that it will be easy to understand for anyone who wants to learn. The link to the websites I scraped is in the URL of every file. Take a look and feel free to ask any questions ;)